data: Move selected hunk recalculation
Hunk navigation now uses focused data modules for selected-change loading,
selected-hunk filtering, and consumed replacement-mask lookup.

The project still keeps post-action hunk recalculation inside the
navigation module, so the module mixes iteration with refresh logic for a
single file after line-level actions.

This commit begins separating that refresh path by moving the selected-
hunk recalculation result enum and file refresh routine into
data.selected_change.hunk_recalculation, then pointing the command refresh
helper at the new data owner.

Subsequent commits will move the matching tests and add an import- boundary
guard for the new owner.
